mod/#60: presentation/home API(조회,좋아요)

+++ app/src/main/java/org/ssg_tab/presentation/ui/home/state/HomeContract.kt
@@ -9,6 +9,6 @@ data class HomeContract(
     val homeFeed: HomeFeedEntity? = null,
     val isLoading: Boolean = false,
     val error: String? = null,
-    val likedContentIds: Set<Int> = emptySet(),
-    val isLikingContent: Set<Int> = emptySet()
+    val likedContentIds: Set<Long> = emptySet(),  // Int -> Long
+    val isLikingContent: Set<Long> = emptySet(),  // Int -> Long
 )
